Fixes for Windows LLP64 in debug output

+++ wavefront/wavefront_backtrace_buffer.c
@@ -510,9 +510,9 @@ bt_block_idx_t wf_backtrace_buffer_compact_marked(
   bt_buffer->num_compactions++;
   // DEBUG
   if (verbose >= 3) {
-    fprintf(stderr,"[WFA::BacktraceBuffer] Compacted from %lu MB to %lu MB (%2.2f%%)",
-        CONVERT_B_TO_MB(read_global_pos*sizeof(bt_block_t)),
-        CONVERT_B_TO_MB(write_global_pos*sizeof(bt_block_t)),
+    fprintf(stderr,"[WFA::BacktraceBuffer] Compacted from %" PRIu64 " MB to %" PRIu64 " MB (%2.2f%%)",
+        (uint64_t)CONVERT_B_TO_MB((uint64_t)read_global_pos*sizeof(bt_block_t)),
+        (uint64_t)CONVERT_B_TO_MB((uint64_t)write_global_pos*sizeof(bt_block_t)),
         100.0f*(float)write_global_pos/(float)read_global_pos);
   }
   // Return last index
